+++ rasa_nlu/converters.py
@@ -161,15 +161,22 @@ def load_wit_data(filename):
     return TrainingData(training_examples)
 
 
-def load_markdown_data(filename):
-    # type: (Text) -> TrainingData
+def load_markdown_data(filenames):
+    # type: (List[Text]) -> TrainingData
     """Loads training data stored in markdown data format."""
     from rasa_nlu.utils.md_to_json import MarkdownToJson
-    data = MarkdownToJson(filename).load()
-    data = data["rasa_nlu_data"]
-    return TrainingData(data["common_examples"],
-                        get_entity_synonyms_dict(data["entity_synonyms"]),
-                        data["regex_features"])
+
+    common_examples = []
+    known_synonyms = {}
+    regex_features = []
+    for filename in filenames:
+        data = MarkdownToJson(filename).load()
+        data = data["rasa_nlu_data"]
+        common_examples.extend(data["common_examples"])
+        known_synonyms = get_entity_synonyms_dict(data["entity_synonyms"],
+                                                  known_synonyms)
+        regex_features.extend(data["regex_features"])
+    return TrainingData(common_examples, known_synonyms, regex_features)
 
 
 def rasa_nlu_data_schema():
@@ -249,26 +256,32 @@ def validate_rasa_nlu_data(data):
         raise e
 
 
-def load_rasa_data(filename):
-    # type: (Text) -> TrainingData
+def load_rasa_data(filenames):
+    # type: (List[Text]) -> TrainingData
     """Loads training data stored in the rasa NLU data format."""
 
-    data = _read_json_from_file(filename)
-    validate_rasa_nlu_data(data)
+    common = list()
+    intent = list()
+    entity = list()
+    regex_features = list()
+    synonyms = list()
+    for filename in filenames:
+        data = _read_json_from_file(filename)
+        validate_rasa_nlu_data(data)
 
-    common = data['rasa_nlu_data'].get("common_examples", list())
-    intent = data['rasa_nlu_data'].get("intent_examples", list())
-    entity = data['rasa_nlu_data'].get("entity_examples", list())
-    regex_features = data['rasa_nlu_data'].get("regex_features", list())
-    synonyms = data['rasa_nlu_data'].get("entity_synonyms", list())
+        common += data['rasa_nlu_data'].get("common_examples", list())
+        intent += data['rasa_nlu_data'].get("intent_examples", list())
+        entity += data['rasa_nlu_data'].get("entity_examples", list())
+        regex_features += data['rasa_nlu_data'].get("regex_features", list())
+        synonyms += data['rasa_nlu_data'].get("entity_synonyms", list())
 
     entity_synonyms = get_entity_synonyms_dict(synonyms)
 
     if intent or entity:
-        logger.warn("DEPRECATION warning: Data file contains 'intent_examples' "
+        logger.warn("DEPRECATION warning: Data file \"{}\" contains 'intent_examples' "
                     "or 'entity_examples' which will be "
                     "removed in the future. Consider putting all your examples "
-                    "into the 'common_examples' section.")
+                    "into the 'common_examples' section.".format(filename))
 
     all_examples = common + intent + entity
     training_examples = []
@@ -281,10 +294,10 @@ def load_rasa_data(filename):
     return TrainingData(training_examples, entity_synonyms, regex_features)
 
 
-def get_entity_synonyms_dict(synonyms):
+def get_entity_synonyms_dict(synonyms, known_synonyms=None):
     # type: (List[Dict]) -> Dict
     """build entity_synonyms dictionary"""
-    entity_synonyms = {}
+    entity_synonyms = {} if not known_synonyms else known_synonyms
     for s in synonyms:
         if "value" in s and "synonyms" in s:
             for synonym in s["synonyms"]:
@@ -352,9 +365,9 @@ def load_data(resource_name, language='en', fformat=None):
     elif fformat == DIALOGFLOW_FILE_FORMAT:
         return load_dialogflow_data(files, language)
     elif fformat == RASA_FILE_FORMAT:
-        return load_rasa_data(files[0])
+        return load_rasa_data(files)
     elif fformat == MARKDOWN_FILE_FORMAT:
-        return load_markdown_data(files[0])
+        return load_markdown_data(files)
     else:
         raise ValueError("unknown training file format : {} for "
                          "file {}".format(fformat, resource_name))