+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Added
 - evaluation of intent classification with crossvalidation in ``evaluation.py``
 - support for inline entity synonyms in markdown training format
 - support for regex features in markdown training format
+- support for splitting training data into multiple files (markdown and JSON only)
 
 Changed
 -------